# manager-external-client/api/openapi.yaml
openapi: 3.0.0
info:
  description: Public API for the Preservation Manager.
  title: Preservation Manager Public API
  version: 1.0.0
servers:
- description: Local server
  url: http://localhost:8484/api
security:
- ApiKeyAuth: []
...
paths:
...
  /{orgName}/job:
    get:
      operationId: listJobs
      parameters:
      - description: The ID of the organization
        example: organization-id
        in: path
        name: orgName
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJobsResponse'
          description: List of jobs
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Lists jobs
      tags:
      - PreservationManager
      x-accepts: application/json
  /job/{jobId}:
    get:
      operationId: describeJob
      parameters:
      - description: The ID of the job
        example: 10293
        in: path
        name: jobId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescribeJobResponse'
          description: Job details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Gets job details
      tags:
      - PreservationManager
      x-accepts: application/json
  /job/validate:
    post:
      operationId: validateObject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidatePreservationObjectRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidatePreservationObjectResponse'
          description: Validate object response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Validates the local copy of an object
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/validateRemote:
    post:
      operationId: validateObjectRemote
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateObjectRemoteRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateObjectRemoteResponse'
          description: Validate object response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Validates a remote copy of an object
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/replicate:
    post:
      operationId: replicateObject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplicateObjectRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplicateObjectResponse'
          description: Replicate object response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Replicates an object from local storage to a remote
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/restore:
    post:
      operationId: restoreObject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestorePreservationObjectRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestorePreservationObjectResponse'
          description: Restore object response
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Restores the local copy of an object from a remote source
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/executing:
    get:
      operationId: listExecutingJobs
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListJobsResponse'
          description: List of jobs
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Lists jobs that have a state of EXECUTING
      tags:
      - PreservationManager
      x-accepts: application/json
  /job/fail:
    post:
      operationId: failJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelJobRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Marks an executing job as failed. This action can only be completed
        by a service administrator, and should only be used when a job has a state
        of executing but is not being executed by a worker.
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/cancel:
    post:
      operationId: cancelJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelJobRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Cancels a pending job
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/disable:
    post:
      operationId: disableJobTypes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisableJobTypesRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Disables job types from being processed
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/enable:
    post:
      operationId: enableJobTypes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableJobTypesRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Enables job types from being processed
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /job/{jobId}/logs:
    get:
      operationId: retrieveJobLogs
      parameters:
      - description: The ID of the job
        example: 10293
        in: path
        name: jobId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveLogsResponse'
          description: Ingest logs
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves all of the logs associated with a job
      tags:
      - PreservationManager
      x-accepts: application/json
...
components:
  schemas:
...
    RestorePreservationObjectRequest:
      example:
        versions:
        - v2
        - v2
        externalObjectId: 1711.dl/ZQ5WPGU2GKBFT8Q
        vault: vault-name
      properties:
        vault:
          description: The name of the vault
          example: vault-name
          type: string
        externalObjectId:
          description: The external ID of the object
          example: 1711.dl/ZQ5WPGU2GKBFT8Q
          type: string
        versions:
          description: Optional, versions to restore
          items:
            $ref: '#/components/schemas/PersistenceVersion'
          type: array
      required:
      - externalObjectId
      - vault
    RestorePreservationObjectResponse:
      example:
        jobId: 10293
      properties:
        jobId:
          description: The id of a job
          example: 10293
          format: int64
          type: integer
    ValidatePreservationObjectRequest:
      example:
        contentFixityCheck: true
        externalObjectId: 1711.dl/ZQ5WPGU2GKBFT8Q
        vault: vault-name
      properties:
        vault:
          description: The name of the vault
          example: vault-name
          type: string
        externalObjectId:
          description: The external ID of the object
          example: 1711.dl/ZQ5WPGU2GKBFT8Q
          type: string
        contentFixityCheck:
          description: True if the digests of all files in the ocfl object should
            be validated
          example: true
          type: boolean
      required:
      - contentFixityCheck
      - externalObjectId
      - vault
    ValidatePreservationObjectResponse:
      example:
        jobId: 10293
      properties:
        jobId:
          description: The id of a job
          example: 10293
          format: int64
          type: integer
    ValidateObjectRemoteRequest:
      example:
        versions:
        - v2
        - v2
        externalObjectId: 1711.dl/ZQ5WPGU2GKBFT8Q
        vault: vault-name
      properties:
        vault:
          description: The name of the vault
          example: vault-name
          type: string
        externalObjectId:
          description: The external ID of the object
          example: 1711.dl/ZQ5WPGU2GKBFT8Q
          type: string
        dataStore:
          $ref: '#/components/schemas/DataStore'
        versions:
          description: Optional, versions to restore. If not specified all versions
            are validated
          items:
            $ref: '#/components/schemas/PersistenceVersion'
          type: array
      required:
      - dataStore
      - externalObjectId
      - vault
    ValidateObjectRemoteResponse:
      example:
        jobIds:
        - 10293
        - 10293
      properties:
        jobIds:
          description: List of job ids for the created validation jobs
          items:
            $ref: '#/components/schemas/JobId'
          type: array
    ReplicateObjectRequest:
      example:
        versions:
        - v2
        - v2
        externalObjectId: 1711.dl/ZQ5WPGU2GKBFT8Q
        vault: vault-name
      properties:
        vault:
          description: The name of the vault
          example: vault-name
          type: string
        externalObjectId:
          description: The external ID of the object
          example: 1711.dl/ZQ5WPGU2GKBFT8Q
          type: string
        destination:
          $ref: '#/components/schemas/DataStore'
        versions:
          description: Optional, versions to restore. If not specified all versions
            are replicated
          items:
            $ref: '#/components/schemas/PersistenceVersion'
          type: array
      required:
      - destination
      - externalObjectId
      - vault
    ReplicateObjectResponse:
      example:
        jobIds:
        - 10293
        - 10293
      properties:
        jobIds:
          description: List of job ids for the created replication jobs
          items:
            $ref: '#/components/schemas/JobId'
          type: array
...
    RetrieveLogsResponse:
      example:
        logEntries:
        - createdTimestamp: 2000-01-23T04:56:07.000+00:00
          message: Something went wrong
        - createdTimestamp: 2000-01-23T04:56:07.000+00:00
          message: Something went wrong
      properties:
        logEntries:
          description: Log entries
          items:
            $ref: '#/components/schemas/LogEntry'
          type: array
...
    CancelJobRequest:
      example:
        jobId: 10293
      properties:
        jobId:
          description: The id of a job
          example: 10293
          format: int64
          type: integer
      required:
      - jobId
    ListJobsResponse:
      example:
        jobs:
        - jobId: 10293
          receivedTimestamp: 2000-01-23T04:56:07.000+00:00
          updatedTimestamp: 2000-01-23T04:56:07.000+00:00
        - jobId: 10293
          receivedTimestamp: 2000-01-23T04:56:07.000+00:00
          updatedTimestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        jobs:
          description: List of jobs
          items:
            $ref: '#/components/schemas/JobDetails'
          type: array
      required:
      - jobs
    DescribeJobResponse:
      example:
        job:
          jobId: 10293
          receivedTimestamp: 2000-01-23T04:56:07.000+00:00
          updatedTimestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        job:
          $ref: '#/components/schemas/JobDetails'
      required:
      - job
...
    DisableJobTypesRequest:
      example:
        types:
        - null
        - null
      properties:
        types:
          description: Job types to disable or an empty list to disable all types
          items:
            $ref: '#/components/schemas/JobType'
          type: array
    EnableJobTypesRequest:
      example:
        types:
        - null
        - null
      properties:
        types:
          description: Job types to enable or an empty list to enable all types
          items:
            $ref: '#/components/schemas/JobType'
          type: array
    ErrorResponse:
      properties:
        timestamp:
          description: Timestamp when the error occurred
          format: date-time
          type: string
        requestId:
          description: UUID assigned to the request
          example: 6037de0a4b2e4611956da540938a00ff
          type: string
        status:
          description: HTTP status code
          example: 500
          format: int32
          type: integer
        error:
          description: Application classification of the error
          type: string
        message:
          description: Description of the error
          example: Invalid input
          type: string
        path:
          description: Request path
          example: /api/ingest/bag
          type: string
      required:
      - error
      - message
      - path
      - requestId
      - status
      - timestamp
...
    DataStore:
      description: A data store location
      enum:
      - IBM_COS
      - GLACIER
      type: string
    PersistenceVersion:
      description: The version identifier of the object version
      example: v2
      type: string
    JobId:
      description: The id of a job
      example: 10293
      format: int64
      type: integer
...
    LogEntry:
      example:
        createdTimestamp: 2000-01-23T04:56:07.000+00:00
        message: Something went wrong
      properties:
        level:
          $ref: '#/components/schemas/LogLevel'
        message:
          description: Log entry text
          example: Something went wrong
          type: string
        createdTimestamp:
          description: The timestamp the log entry was created
          format: date-time
          type: string
    LogLevel:
      description: Log level
      enum:
      - INFO
      - WARN
      - ERROR
      type: string
...
    JobDetails:
      example:
        jobId: 10293
        receivedTimestamp: 2000-01-23T04:56:07.000+00:00
        updatedTimestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        jobId:
          description: The id of a job
          example: 10293
          format: int64
          type: integer
        type:
          $ref: '#/components/schemas/JobType'
        state:
          $ref: '#/components/schemas/JobState'
        receivedTimestamp:
          description: Timestamp the job was created
          format: date-time
          type: string
        updatedTimestamp:
          description: Timestamp the job was updated
          format: date-time
          type: string
    JobType:
      description: Type of job
      enum:
      - RETRIEVE_OBJECTS
      - REPLICATE
      - RESTORE
      - FINALIZE_RESTORE
      - VALIDATE_LOCAL
      - PROCESS_BATCH
      - VALIDATE_REMOTE
      - DELETE_DIP
      - CLEANUP_SIPS
      type: string
    JobState:
      description: The state of a job
      enum:
      - PENDING
      - EXECUTING
      - FAILED
      - COMPLETE
      - CANCELLED
      type: string
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-KEY
      type: apiKey